+++ patches/src/main/kotlin/app/morphe/patches/reddit/customclients/SpoofClientPatch.kt
@@ -2,6 +2,7 @@ package app.morphe.patches.reddit.customclients
 
 import app.morphe.patcher.patch.BytecodePatchBuilder
 import app.morphe.patcher.patch.Option
+import app.morphe.patcher.patch.PatchException
 import app.morphe.patcher.patch.bytecodePatch
 import app.morphe.patcher.patch.stringOption
 
@@ -42,56 +43,44 @@ fun spoofClientPatch(
     block: BytecodePatchBuilder.(
         clientIdOption: Option<String>,
         redirectUriOption: Option<String>,
+        userAgentOption: Option<String>
     ) -> Unit,
 ) = bytecodePatch(
     name = "Spoof client",
     description = "Restores functionality of the app by using custom client ID.",
 ) {
-    block(
-        stringOption(
-            "client-id",
-            null,
-            null,
-            "OAuth client ID",
-            "The Reddit OAuth client ID. " +
-                "You can get your client ID from https://www.reddit.com/prefs/apps. " +
-                "The application type has to be \"Installed app\" and the redirect " +
-                "URI has to match the value provided for the \"Redirect URI\" option.",
-            true,
-        ),
-        stringOption(
-            "redirect-uri",
-            "http://127.0.0.1:8080",
-            null,
-            "Redirect URI",
-            "The Reddit OAuth redirect URI. Should be a valid URI.",
-            true,
-        ),
+    val clientIdOption = stringOption(
+        "client-id",
+        null,
+        null,
+        "OAuth client ID",
+        "The Reddit OAuth client ID. Refer to Patcheddit documentation for " +
+                "more information on what to put here. An empty value keeps the app's " +
+                "existing client ID.",
+        false,
     )
-}
-
-/**
- * Base class for patches that spoof the Reddit user agent.
- *
- * @param block The patch block. It is called with the user agent option.
- */
-fun spoofUserAgentPatch(
-    block: BytecodePatchBuilder.(
-        userAgentOption: Option<String>,
-    ) -> Unit,
-) = bytecodePatch(
-    name = "Spoof user agent",
-    description = "Restores functionality of the app by using custom user agent.",
-) {
-    block(
-        stringOption(
+    val redirectUriOption = stringOption(
+        "redirect-uri",
+        null,
+        null,
+        "Redirect URI",
+        "The Reddit OAuth redirect URI. Refer to Patcheddit documentation for " +
+                "more information on what to put here. An empty value keeps the app's " +
+                "existing redirect URI.",
+        false,
+    )
+    val userAgentOption = stringOption(
         "user-agent",
         null,
         null,
         "User agent",
-        "The app's user agent. User agent should be in the format " +
-                "\"<platform>:<app id>:<version> (by /u/<username>)\".",
-        true,
-        )
+        "The app's user agent. Refer to Patcheddit documentation for " +
+                "more information on what to put here. An empty value keeps the app's existing " +
+                "user agent.",
+        false
+    )
+
+    block(
+        clientIdOption, redirectUriOption, userAgentOption
     )
 }