+++ patches/src/main/kotlin/app/morphe/patches/reddit/customclients/boostforreddit/api/SpoofClientPatch.kt
@@ -3,8 +3,8 @@ package app.morphe.patches.reddit.customclients.boostforreddit.api
 import app.morphe.patcher.extensions.InstructionExtensions.addInstructions
 import app.morphe.patcher.extensions.InstructionExtensions.getInstruction
 import app.morphe.patcher.extensions.InstructionExtensions.replaceInstruction
+import app.morphe.patcher.patch.PatchException
 import app.morphe.patches.reddit.customclients.spoofClientPatch
-import app.morphe.patches.reddit.customclients.spoofUserAgentPatch
 import app.morphe.util.getReference
 import app.morphe.util.indexOfFirstInstructionOrThrow
 import app.morphe.util.returnEarly
@@ -14,66 +14,75 @@ import com.android.tools.smali.dexlib2.iface.reference.TypeReference
 
 const val JRAW_NEW_URL_EXTENSION_CLASS_DESCRIPTOR = "Lapp/morphe/extension/boostforreddit/http/HttpUtils;"
 
-val spoofClientPatch = spoofClientPatch { clientIdOption, redirectUriOption ->
-    compatibleWith("com.rubenmayayo.reddit")
+val spoofClientPatch = spoofClientPatch { clientIdOption, redirectUriOption, userAgentOption ->
+    compatibleWith("com.rubenmayayo.reddit"("1.12.12"))
 
     val clientId by clientIdOption
+    val redirectUri by redirectUriOption
+    val userAgent by userAgentOption
 
     execute {
-        // region Patch client id and redirect URI.
-
-        getClientIdFingerprint.method.returnEarly(clientId!!)
-
-        listOf(loginActivityOnCreateFingerprint, loginActivityAShouldOverrideUrlLoadingFingerprint).forEach { fingerprint ->
-            fingerprint.method.let {
-                fingerprint.stringMatches!!.forEach { match ->
-                    val register = it.getInstruction<OneRegisterInstruction>(match.index).registerA
-                    it.replaceInstruction(match.index, "const-string v$register, \"$redirectUriOption\"")
-                }
-            }
+        if (clientIdOption.value == null && redirectUriOption.value == null && userAgentOption.value == null) {
+            throw PatchException("When spoofing client, at least one of clientId, redirectUri or userAgent should be set.")
         }
 
-        loginActivityAShouldOverrideUrlLoadingFingerprint.method.apply {
-            val index = indexOfFirstInstructionOrThrow {
-                opcode == Opcode.IF_EQZ
+        // region Patch client id.
+
+        if (clientId != null) {
+            getClientIdFingerprint.method.returnEarly(clientId!!)
+        }
+
+        // endregion
+
+        // region Patch redirect URI.
+        if (redirectUri != null) {
+            listOf(loginActivityOnCreateFingerprint, loginActivityAShouldOverrideUrlLoadingFingerprint).forEach { fingerprint ->
+                fingerprint.method.let {
+                    fingerprint.stringMatches.forEach { match ->
+                        val register = it.getInstruction<OneRegisterInstruction>(match.index).registerA
+                        it.replaceInstruction(match.index, "const-string v$register, \"$redirectUriOption\"")
+                    }
+                }
             }
-            addInstructions(
-                index + 1,
-                """
+
+            loginActivityAShouldOverrideUrlLoadingFingerprint.method.apply {
+                val index = indexOfFirstInstructionOrThrow {
+                    opcode == Opcode.IF_EQZ
+                }
+                addInstructions(
+                    index + 1,
+                    """
                 const-string v1, "$redirectUriOption"
                 const-string v2, "http://localhost"
                 invoke-virtual {v6, v1, v2}, Ljava/lang/String;->replace(Ljava/lang/CharSequence;Ljava/lang/CharSequence;)Ljava/lang/String;
                 move-result-object v6
                 """
-            )
-        }
-
-        jrawNewUrlFingerprint.method.apply {
-            val index = indexOfFirstInstructionOrThrow {
-                opcode == Opcode.NEW_INSTANCE && getReference<TypeReference>()?.type == "Ljava/net/URL;"
+                )
             }
-            addInstructions(
-                index,
-                """
+
+            jrawNewUrlFingerprint.method.apply {
+                val index = indexOfFirstInstructionOrThrow {
+                    opcode == Opcode.NEW_INSTANCE && getReference<TypeReference>()?.type == "Ljava/net/URL;"
+                }
+                addInstructions(
+                    index,
+                    """
                 invoke-static       { p0 }, $JRAW_NEW_URL_EXTENSION_CLASS_DESCRIPTOR->createUrl(Ljava/lang/String;)Ljava/net/URL;
                 move-result-object  v0
                 return-object v0
                 """
-            )
+                )
+            }
         }
-
         // endregion
-    }
-}
 
-val userAgentPatch = spoofUserAgentPatch { userAgentOption ->
-    compatibleWith("com.rubenmayayo.reddit")
-    execute {
-        // region Patch user agent.
-        buildUserAgentFingerprint.method.let {
-            buildUserAgentFingerprint.stringMatches!!.forEach { match ->
-                val register = it.getInstruction<OneRegisterInstruction>(match.index).registerA
-                it.replaceInstruction(match.index, "const-string v$register, \"$userAgentOption\"")
+        // Patch user agent.
+        if (userAgent != null) {
+            buildUserAgentFingerprint.method.let {
+                buildUserAgentFingerprint.stringMatches.forEach { match ->
+                    val register = it.getInstruction<OneRegisterInstruction>(match.index).registerA
+                    it.replaceInstruction(match.index, "const-string v$register, \"$userAgentOption\"")
+                }
             }
         }
         // endregion