+++ patches/src/main/kotlin/app/morphe/patches/reddit/customclients/redditisfun/api/Fingerprints.kt
@@ -31,3 +31,40 @@ internal val getUserAgentFingerprint = Fingerprint(
         Opcode.CONST,
     )
 )
+
+// TODO: These obfuscated names are bad, but the app is abandoned so these will not be changing.
+// This can be fixed later once patcher API is capable of matching multiple methods with one fingerprint.
+internal val oAuth2ActivityD0Fingerprint = Fingerprint(
+    strings = listOf("redditisfun://auth"),
+    custom = { method, classDef ->
+        if (!classDef.endsWith("OAuth2Activity${'$'}b;")) return@Fingerprint false
+
+        method.name == "d0"
+    }
+)
+
+internal val oAuth2ActivityShouldOverrideUrlLoadingFingerprint = Fingerprint(
+    strings = listOf("redditisfun://auth"),
+    custom = { method, classDef ->
+        if (!classDef.endsWith("OAuth2Activity${'$'}a;")) return@Fingerprint false
+
+        method.name == "shouldOverrideUrlLoading"
+    }
+)
+
+internal val cActivityJFingerprint = Fingerprint(
+    strings = listOf("redditisfun://auth"),
+    custom = { method, classDef ->
+        if (!classDef.endsWith("c;")) return@Fingerprint false
+
+        method.name == "j"
+    }
+)
+
+// Should usually match: g2/c.x()
+internal val imgurApiFingerprint = Fingerprint(
+    accessFlags = listOf(AccessFlags.PRIVATE, AccessFlags.STATIC),
+    returnType = "Landroid/net/Uri;",
+    parameters = listOf("Ljava/lang/String;", "Z"),
+    strings = listOf("https", "api", "imgur", "3", "gallery", "album")
+)