+++ patches/src/main/kotlin/app/morphe/patches/reddit/customclients/relayforreddit/api/SpoofClientPatch.kt
@@ -3,30 +3,56 @@ package app.morphe.patches.reddit.customclients.relayforreddit.api
 import app.morphe.patcher.extensions.InstructionExtensions.addInstructions
 import app.morphe.patcher.extensions.InstructionExtensions.getInstruction
 import app.morphe.patcher.extensions.InstructionExtensions.replaceInstruction
+import app.morphe.patcher.patch.PatchException
 import app.morphe.patches.reddit.customclients.spoofClientPatch
 import com.android.tools.smali.dexlib2.Opcode
 import com.android.tools.smali.dexlib2.builder.instruction.BuilderInstruction10t
 import com.android.tools.smali.dexlib2.builder.instruction.BuilderInstruction21t
 import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
 
-val spoofClientPatch = spoofClientPatch(redirectUri = "dbrady://relay") {
+val spoofClientPatch = spoofClientPatch { clientIdOption, redirectUriOption, userAgentOption ->
     compatibleWith(
-        "free.reddit.news",
-        "reddit.news",
+        "free.reddit.news"("10.2.40"),
+        "reddit.news"("10.2.40"),
     )
 
-    val clientId by it
+    val clientId by clientIdOption
+    val redirectUri by redirectUriOption
+    val userAgent by userAgentOption
 
     execute {
-        // region Patch client id.
+        if (clientId == null) {
+            throw PatchException("When spoofing client, clientId must be set.")
+        }
 
+        // region Patch redirect URI.
+        if (redirectUri != null) {
+            setOf(
+                loginActivityRedirectUriFingerprint,
+                shouldOverrideUrlLoadingRedirectUriFingerprint,
+                redditAccountManagerRedirectUriFingerprint
+            ).forEach { fingerprint ->
+                val redirectUriIndex = fingerprint.stringMatches.last().index
+                fingerprint.method.apply {
+                    val redirectUriRegister = getInstruction<OneRegisterInstruction>(redirectUriIndex).registerA
+
+                    fingerprint.method.replaceInstruction(
+                        redirectUriIndex,
+                        "const-string v$redirectUriRegister, \"$redirectUri\"",
+                    )
+                }
+            }
+        }
+        // endregion
+
+        // region Patch client id.
         setOf(
             loginActivityClientIdFingerprint,
             getLoggedInBearerTokenFingerprint,
             getLoggedOutBearerTokenFingerprint,
             getRefreshTokenFingerprint,
         ).forEach { fingerprint ->
-            val clientIdIndex = fingerprint.stringMatches!!.first().index
+            val clientIdIndex = fingerprint.stringMatches.first().index
             fingerprint.method.apply {
                 val clientIdRegister = getInstruction<OneRegisterInstruction>(clientIdIndex).registerA
 
@@ -36,7 +62,24 @@ val spoofClientPatch = spoofClientPatch(redirectUri = "dbrady://relay") {
                 )
             }
         }
+        // endregion
 
+        // region Patch user agent.
+        if (userAgent != null) {
+            networkModuleUserAgentFingerprint.apply {
+                val invokeDirectIndex = instructionMatches.first().index
+                val invokeDirectRegister = method.getInstruction<OneRegisterInstruction>(invokeDirectIndex).registerA
+
+                val userAgentField = classDef.fields.first { field ->
+                    field.type == "Ljava/lang/String;"
+                }
+
+                method.addInstructions(invokeDirectIndex, """
+                    const-string v$invokeDirectRegister, "$userAgent"
+                    sput-object v$invokeDirectRegister, ${userAgentField.definingClass}->${userAgentField.name}:Ljava/lang/String;
+                """)
+            }
+        }
         // endregion
 
         // region Patch miscellaneous.