+++ CHANGELOG.md
@@ -1,3 +1,34 @@
+# [1.1.0-dev.4](https://github.com/wchill/patcheddit/compare/v1.1.0-dev.3...v1.1.0-dev.4) (2026-01-09)
+
+
+### Features
+
+* Add support for patching Joey with custom redirect URI ([bc7bc1a](https://github.com/wchill/patcheddit/commit/bc7bc1a7acf05eacf7fff8f62a217cfdeb027049))
+* Add support for patching Sync with custom redirect URI ([8822a60](https://github.com/wchill/patcheddit/commit/8822a605d4c41cd3ba49fe60f85b50531daff4d5))
+
+# [1.1.0-dev.3](https://github.com/wchill/patcheddit/compare/v1.1.0-dev.2...v1.1.0-dev.3) (2026-01-09)
+
+
+### Features
+
+* Add support for patching BaconReader with custom redirect URI ([a99d598](https://github.com/wchill/patcheddit/commit/a99d5980b22a25843f97d88045874c73435940dd))
+
+# [1.1.0-dev.2](https://github.com/wchill/patcheddit/compare/v1.1.0-dev.1...v1.1.0-dev.2) (2026-01-09)
+
+
+### Features
+
+* Add support for patching Relay with custom redirect URI ([55a1958](https://github.com/wchill/patcheddit/commit/55a1958008086651e7d97a1588c728325a08a7aa))
+
+# [1.1.0-dev.1](https://github.com/wchill/patcheddit/compare/v1.0.1...v1.1.0-dev.1) (2026-01-09)
+
+
+### Features
+
+* Add imgur album patch for rif (from null-dev/revanced-patches@1dabe26) ([d7cc5a1](https://github.com/wchill/patcheddit/commit/d7cc5a1a653991c759d53171b09a6b6a69c08518))
+* Add rif premium unlock patch (from null-dev/revanced-patches@1dabe26) ([aff0385](https://github.com/wchill/patcheddit/commit/aff0385b158f883b2bbd445a16b1d9c5b5175b2d))
+* Add support for patching rif with custom redirect URI (from altherat/revanced-patches@fe01d43) ([e2aa03a](https://github.com/wchill/patcheddit/commit/e2aa03ae1bee92a499166647e9d19e5492904a8d))
+
 # 1.0.0 (2026-01-08)